Mark map tasks a complete and detect job completion
When jobs are complete, the CompleteTask RPC is called and the coordinator is signaled of the completed task via the completeCh channel. The coordinator keeps track of the number of completed jobs and when all of the tasks are complete, marks done on the doneCh channel. The Done method uses a non-blocking select to read from the doneCh to determine if the coordinator is sone.

Not yet supported is task timeouts.