Add task timeout and retries
When a task is requested a goroutine is also started that will signal a timeout.

When a task is completed, its status is set to Complete. When a timeout or complete for a task is received, the status is checked and the event will be ignored if the task was marked as completed.