Implement happy path AppendEntries and commit
The log is currently a zero based index log, while the test system expects an 1-based index log. To work around this, Start returns the index + 1 and the same is done when sending a message to applyChn.

This change doesn't handle AppendEntries that fail due to log inconsistencies and decrementing nextIndex yet.

It also doesn't implement election validation (section 5.4)