Fix bug - assertion that applyLog should only be called after commitIndex update
Fix for a subtle timing back in an assertion of the applyLog logic. The check will fail when commitIndex and lastApplied are the same. Since applyLog is called in its own goroutinue, it's possible that multiple invocations are queued up between successive updates to commitIndex as multiple followers respond to AppendEntries with different index locations.