Fix endless elections bug when running TestManyElections2B
When only a minimum majority of nodes are still up, it's possible for two nodes to have the same election timeout. At that point, they will stay in lockstep and neither will win an election.

The fix is to add some jitter after several failed campaign attempts.

Interesting, the test was failing ~ 1 / 300 tests and the original random election was over 300 millisecond range.